requirements.txt
streamlit>=1.31.0
crewai>=0.28.8
langchain>=0.1.10
langchain-openai>=0.0.8
langchain-community>=0.0.25
openai>=1.12.0
pydantic>=2.6.0,<3.0.0
python-dotenv>=1.0.0,<1.1.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
pandas>=2.2.0
plotly>=5.18.0
markdown>=3.5.0
PyYAML>=6.0.0
tenacity>=8.2.0 